feat: fetch snowflake_password in config object
